[Mod] extension

+++ db/schema.rb
@@ -85,8 +85,8 @@ ActiveRecord::Schema.define(version: 2023_09_21_110756) do
     t.string "name", null: false
     t.text "body", null: false
     t.string "address", null: false
-    t.decimal "latitude", null: false
-    t.decimal "longitude", null: false
+    t.float "latitude", null: false
+    t.float "longitude", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["user_id"], name: "index_posts_on_user_id"
